requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

diffusers==0.30.2
langchain-core==0.3.5
langchain-community==0.3.0
langchain-openai==0.2.0
matplotlib==3.9.0
numpy==1.26.4
openai==1.35.4
pandas==2.2.2
pillow==10.3.0
seaborn==0.13.2
torch==2.3.1+cu118
tqdm==4.66.4
transformers==4.41.2
